% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addLemmaToCDB}
\alias{addLemmaToCDB}
\title{Add a LEMMA table to a concept database (CDB)}
\usage{
addLemmaToCDB(
  CDB,
  tablenames = c("FINDINGS", "QUAL", "BODY", "LATERALITY", "SEVERITY", "CAUSES", "MORPH"),
  min_len_normalize = 3
)
}
\arguments{
\item{CDB}{concept database environment}

\item{tablenames}{which data tables to add lemmatized terms}

\item{min_len_normalize}{tokens with fewer than this number of
characters will not be lemmatised (default 3)}
}
\value{
CDB environment with an additional 'LEMMA' table. If it
  already exists it will be over-written
}
\description{
Adds a 'LEMMA' data.table containing the original term and lemmatized
version for fast concept lookup. The function uses the spacy
lemmatizer. The function does not
lemmatize terms which are partially or fully capitalised
as these need to be matched in a case sensitive manner.
}
\examples{
# Not run
# CDB <- addLemmaToCDB(CDB)
}
\seealso{
findConceptMatch
}
