% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluateNER}
\alias{evaluateNER}
\title{Evaluate the performance of named entity recognition compared to
a gold standard}
\usage{
evaluateNER(
  actual,
  goldstandard,
  relaxed = FALSE,
  aggregate = TRUE,
  SNOMED = Rdiagnosislist::getSNOMED(),
  subset = NULL
)
}
\arguments{
\item{actual}{data.frame containing columns id (document identifier)
and conceptId, e.g. output of NERcorpus}

\item{goldstandard}{data.frame containing columns id (document
identifier) and conceptId (character, which can contain multiple
concepts separated by |). If the conceptId column is integer64, it
is assumed that there is only one correct SNOMED CT concept for
each mention.}

\item{relaxed}{Boolean TRUE or FALSE. TRUE means that any ancestor
of a gold standard concept is considered correct. FALSE means that
only exact concepts are counted}

\item{aggregate}{Boolean TRUE or FALSE. TRUE means that all counts
aggregated across all documents before calculating precision,
recall and F1. FALSE returns results per document}

\item{SNOMED}{environment containing a SNOMED dictionary}

\item{subset}{SNOMEDconcept vector specifying which concepts
to include in this evaluation (e.g. only disorders)}
}
\value{
data.table with columns precision, recall, F1, and also
  an id column if aggregate = FALSE.
}
\description{
Compares the extraction of diagnosis data to a gold standard
in the following way. Firstly, the comparison considers only unique
mentions of a concept (repeats are ignored). Secondly, the comparison
ignores ancestors of any concept. For example, if a document contains
both 'diabetes mellitus' (ancestor) and 'diabetic retinopathy'
(descendant which implies the presence of diabetes mellitus), the
results will be the same whether or not the diabetes mellitus concept
is returned by the algorithm.
}
\examples{
require(Rdiagnosislist)
require(data.table)
SNOMED <- sampleSNOMED()

actual = data.table(id = c(2, 2, 3, 3, 3),
  conceptId = as.SNOMEDconcept(c(
  'Diastolic heart failure', 
  'Systolic heart failure', 
  'Systolic dysfunction', 
  'Acute heart failure',
  'Cardiac finding')))
# first two are correct, third concept is vaguer than gold standard
# (i.e. true in 'relaxed' mode but not if relaxed = FALSE),
# fourth concept is a false positive and fifth is an ancestor
# which should be ignored

goldstandard = data.table(id = c(1, 2, 2, 2, 3, 3),
  conceptId = c('703276005', '78862003', '417996009|703273002',
  '443253003', '56675007', '84114007'))
# first and second are not present in actual (i.e. false negatives),
# third has two correct options, last is an ancestor of another
# term so it should be ignored

evaluateNER(actual, goldstandard)
evaluateNER(actual, goldstandard, relaxed = TRUE)
evaluateNER(actual, goldstandard, aggregate = FALSE)
}
\seealso{
NERcorpus
}
