% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllergy.R, R/findAttr.R, R/findBody.R,
%   R/findCause.R, R/findLaterality.R
\name{findAllergy}
\alias{findAllergy}
\alias{findAttr}
\alias{findBody}
\alias{findCause}
\alias{findLaterality}
\title{Find attributes of a concept}
\usage{
findAllergy(D, i)

findAttr(D, i)

findBody(D, i)

findCause(D, i)

findLaterality(D, i)
}
\arguments{
\item{D}{spacy parse table containing columns: token, lemma, dep_rel,
head, semType}

\item{i}{= vector of indices of base word(s)}
}
\value{
integer vector of indices of linked words, or a zero length
  vector if no link
}
\description{
Finds concepts semantically related to the index concept.
For findAllergy, the semantic type is limited to substances.
For findBody, the semantic type is limited to body structures or
  laterality.
For findLaterality, the semantic type is limited to laterality.
For findCause, the semantic type must be one of finding,
  disorder, organism, substance, event or procedure.
}
\examples{
D <- showparse("ulcers of the left big and second toes due to diabetes")
findAttr(D, 4)
findBody(D, 4)
findCause(D, 4)
findAllergy(D, 4)
}
\seealso{
addAnd, findAllergy, findAttr, findBody, findCause,
  findLaterality, linkFrom, linkTo

Other linkFunction: 
\code{\link{addAnd}()},
\code{\link{linkFrom}()}

Other linkFunction: 
\code{\link{addAnd}()},
\code{\link{linkFrom}()}

Other linkFunction: 
\code{\link{addAnd}()},
\code{\link{linkFrom}()}

Other linkFunction: 
\code{\link{addAnd}()},
\code{\link{linkFrom}()}

Other linkFunction: 
\code{\link{addAnd}()},
\code{\link{linkFrom}()}
}
\concept{linkFunction}
