% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parseSentence}
\alias{parseSentence}
\title{Parse a sentence using spacy and find annotations}
\usage{
parseSentence(text, CDB, SNOMED, wordlimit = 6, keep_all_matches = FALSE)
}
\arguments{
\item{text}{text phrase to match}

\item{CDB}{concept database environment}

\item{SNOMED}{SNOMED CT dictionary environment}

\item{wordlimit}{maximum number of words in a phrase to attempt
to match to a SNOMED concept}
}
\value{
spacy parse object (output of showparse) with an attribute
 'annotations' which is a data.table with columns conceptId,
 startword (index of first word of matched phrase),
 endword (index of last word of matched phrase),
 startwhole (index of first word of entire concept, in this case
 same as startword), endwhole (index of last word of entire concept,
 in this case same as endword), semType (semantic type of
 matched SNOMED CT concept), due_to (linked cause concept),
 causing (linked consequence concept), attributes (list of 
 concept attributes), link_to (concept that this concept is
 an attribute of), laterality (SNOMED CT concept for laterality of
 of concept
}
\description{
Parses a sentence using showparse, and finds annotations using
findConceptMatch.
}
\seealso{
showparse
}
