% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testlink.R
\docType{data}
\name{testlink_data}
\alias{testlink_data}
\title{Test file for linking functions}
\format{
An object of class \code{"data.table"} and
  \code{"data.frame"}
}
\usage{
data(testlink_data)
}
\description{
A table of tests of link functions
}
\details{
\describe{
  \item{linkFunction}{ character: function to test; one of
    addAnd, findLaterality, findAttr, findBody, findCause,
    or findAllergy}
  \item{text}{ character: text to test}
  \item{semTypes}{ character: string with number of characters
    equal to number of words in text,
    containing letters that represent the semantic type of a word:
    (space) = (any semantic type), F = finding, D = disorder,
    B = body structure, L = laterality, O = organism, S = substance,
    P = procedure, Z = severity, Q = qualifier value}
  \item{index_start}{ integer: position of the first word that
    the link is in relation to}
  \item{index_end}{ integer: position of the last word that the
    link is in relation to; same as index_start if it is a single
    word}
  \item{expected}{ character: correct output in the format:
    "# word, # word2|# alt" where # is the
    index number of the word linked to, word/word2/alt are the words
    linked to, and | separates different options for correct outputs}
}
}
\examples{
# Show properties of the testlink_data table
data('testlink_data')
str(testlink_data)
}
\seealso{
testlink, testlink_from_data
}
\keyword{datasets}
