% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testlink.R
\name{testlink_from_data}
\alias{testlink_from_data}
\alias{testlink}
\title{Test the text linking functions}
\usage{
testlink_from_data(testlink_data = NULL)

testlink(
  linkFunction = c("addAnd", "findLaterality", "findAttr", "findBody", "findCause",
    "findAllergy"),
  text,
  semTypes,
  i,
  expected = NULL
)
}
\arguments{
\item{testlink_data}{data.frame containing link function tests
in the format
one row per test, with columns: linkFunction (link function),
text, semTypes (semantic type of each word in the format of
semTypes as described below), index_start (start position of index
word for linkage,
index_end (end position of index word for linkage, same as 
index_start if a single word), expected
(correct output in the format: "# word, # word2|# alt" where # is the
index number of the word linked to, word/word2/alt are the words
linked to, and | separates different options for correct outputs).
If NULL, the provided dataset testlink_data will be used.}

\item{linkFunction}{a function or character vector of length 1
specifying the name of the link function to test which can be
retrieved with get (one of 'findLaterality', 'findAttr',
'findBody', 'findCause', or 'findAllergy')}

\item{text}{phrase to test}

\item{semTypes}{either a character vector containing the semantic
type of each word in text, or a single string of length 1
and number of characters equal to number of words in text,
containing letters that represent the semantic type of a word:
(space) = (any semantic type), F = finding, D = disorder,
B = body structure, L = laterality, O = organism, S = substance,
P = procedure, Z = severity, Q = qualifier value}

\item{i}{integer vector denoting the index number or range of word
that the link is in relation to}

\item{expected}{character vector where each element is a possible
correct output in the format: "# word, # word2" where # is the
index number of the word linked to, and word/word2 are the words
linked to}
}
\value{
testlink returns a single string equal to TRUE if the
  output is equal to expected, and the actual output
  if there is a difference. testlink_from_data returns a character
  vector containing the result for each row
}
\description{
Test a single example or use a dataset to carry out a series of tests
}
\examples{
data(testlink_data)
testlink_from_data(testlink_data)
}
\seealso{
testlink_data
}
