% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAnd.R
\name{addAnd}
\alias{addAnd}
\title{Link concepts by 'and'}
\usage{
addAnd(D, i)
}
\arguments{
\item{D}{spacy parse table containing columns: token, lemma, dep_rel,
head, semType. Can be created by the 'showparse' function}

\item{i}{= vector of indices of base word(s)}
}
\value{
integer vector of indices of linked words, or a zero length
  vector if no link
}
\description{
Finds concepts semantically related to the index concept using the
conjunction 'and'. This function does not limit the output to
concepts of a particular semantic type, which might have to be 
filtered afterwards.
}
\examples{
D <- showparse("fracture of left femur and right humerus")
addAnd(D, 4)
}
\seealso{
addAnd, findAllergy, findAttr, findBody, findCause,
  findLaterality, linkFrom, linkTo

Other linkFunction: 
\code{\link{findAllergy}()},
\code{\link{linkFrom}()}
}
\concept{linkFunction}
