% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addLemmaToCDB}
\alias{addLemmaToCDB}
\title{Add lemmatized versions of terms to a concept database (CDB)}
\usage{
addLemmaToCDB(
  CDB,
  tablenames = c("FINDINGS", "QUAL", "BODY", "LATERALITY", "SEVERITY", "CAUSES", "MORPH"),
  ignore_if_already = TRUE,
  min_len_normalize = 5
)
}
\arguments{
\item{CDB}{concept database environment}

\item{tablenames}{which data tables to add lemmatized terms}

\item{ignore_if_already}{TRUE or FALSE, whether to skip if
a table already has a lemma column}

\item{min_len_normalize}{tokenswith fewer than this number of
characters will not be lemmatised}
}
\value{
CDB environment with lemma columns added to relevant tables
}
\description{
Adds a 'lemma' column to the relevant tables within the concept
database, using the spacy lemmatizer, in order to assist term
detection. The function does not
lemmatize terms which are partially or fully capitalised
as these need to be matched in a case sensitive manner.
}
\examples{
# Not run
# CDB <- addLemmaToCDB(CDB)
}
\seealso{
findConceptMatch
}
